Fix integer division in YOLO box scale factors.
input.cols/rows and size.width/height are all int, so the scale
factors were computed with integer division and truncated before
being stored, producing wrong (or zero) detection box scaling.

+++ ext/yolo.cpp
@@ -63,8 +63,8 @@ void Yolo::process(cv::Mat &img, std::vector<Yolo::Item> &items) {
     float *data    = (float *)outputs[0].data;
 
     // Resizing factor.
-    float x_factor = input.cols / size.width;
-    float y_factor = input.rows / size.height;
+    float x_factor = (float)input.cols / size.width;
+    float y_factor = (float)input.rows / size.height;
 
     // Post-process
     std::vector<int>       class_ids;
